requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
anthropic==0.68.0
python-dotenv==1.0.0
schedule==1.2.0
geopy==2.4.0
folium==0.15.0
python-socketio==5.9.0
eventlet==0.33.3
httpx==0.25.2
